moved config sub-classes into separate files
